requirements.txt
asgiref==3.8.1
bcc==0.1.10
Django==5.2.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
numpy==2.2.5
packaging==25.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
traitlets==5.14.3
traittypes==0.2.1
uritemplate==4.1.1
